feat: modify data type

See merge request s13-fintech-finance-sub1/S13P21A407!74

+++ pocketc_backend/src/main/java/com/ssafy/pocketc_backend/domain/user/dto/request/UserUpdateRequest.java
@@ -5,5 +5,5 @@ public record UserUpdateRequest(
     String name,
     String email,
     String password,
-    Integer budget
+    Long budget
 ) {}
